Squash GHA tests fix
