--- server_src/routes/history.js.bak
@@ -1,24 +1,0 @@
-
-// router.get("/info/", ensureAuthenticated, function(req, res) {
-//   //req.socket.setTimeout(Number.MAX_VALUE);
-//   res.writeHead(200, {
-//     "Content-Type": "text/event-stream", // <- Important headers
-//     "Cache-Control": "no-cache",
-//     Connection: "keep-alive"
-//   });
-//   res.write("\n");
-//   (function(clientId) {
-//     clients[clientId] = res; // <- Add this client to those we consider "attached"
-//     req.on("close", function() {
-//       delete clients[clientId];
-//     }); // <- Remove this client when he disconnects
-//   })(++clientId);
-//   //console.log("Client: " + Object.keys(clients));
-// });
-// setInterval(async function() {
-//   for (clientId in clients) {
-//     for (clientId in clients) {
-//       clients[clientId].write("data: " + historyTable + "\n\n"); // <- Push a message to a single attached client
-//     }
-//   }
-// }, 10000);
